refactor(flui-painting): Mythos Step 4 -- split canvas.rs god module
Split the 3,305-LOC `crates/flui-painting/src/canvas.rs` god module
into a concern-based `canvas/` directory. Eight files, each focused
on one concern; total LOC drops to ~2,200 (net -1,100 from removing
duplicate docstring noise + intermediate boilerplate).

Files created:

| File                              | LOC | Concern                                             |
|-----------------------------------|-----|-----------------------------------------------------|
| `canvas/mod.rs`                   | 350 | Canvas struct + Default + AsRef + lifecycle + queries + add_hit_region |
| `canvas/state.rs`                 | 188 | CanvasState + ClipShape + save/restore + save_layer family |
| `canvas/transform.rs`             |  94 | translate/scale_*/rotate*/skew/transform/set_transform/transform_matrix |
| `canvas/clipping.rs`              | 148 | clip_rect/clip_rrect/clip_path/*_ext + bounds queries |
| `canvas/drawing.rs`               | 456 | 29 primary draw_* primitives (each emits one DrawCommand) |
| `canvas/scoped.rs`                | 174 | 12 with_* closure-based scoped helpers              |
| `canvas/composition.rs`           | 114 | extend_from/extend/merge/append_* + record/build    |
| `canvas/sugar.rs`                 | 663 | chaining API + batch + convenience shapes + debug + conditional + grid/repeat |

The split honors concerns by data shape and call site, not by file
size. Each file owns one invariant:

- state.rs owns the save/restore stack and its invariants
  (`save()` push, `restore()` pop, depth markers for clip truncation).
- transform.rs owns the transform matrix mutation surface.
- clipping.rs owns the clip stack push and the corresponding
  `DrawCommand::Clip*` emission.
- drawing.rs owns the 29 primary `DrawCommand`-emitting methods --
  this file is the largest because each method is structurally
  similar (pattern match → `display_list.push(...)`).
- scoped.rs owns the closure-based save/restore wrappers (zero-cost;
  compile to direct save+restore sequences).
- composition.rs owns multi-canvas merge + static constructors
  (`Canvas::record`, `Canvas::build`).
- sugar.rs owns caller-side ergonomic wrappers (chaining API, batch
  loops, convenience shapes, debug visualization, conditional draws,
  grid/repeat patterns).

The 7 inline `#[cfg(test)] mod tests` from the old canvas.rs are
preserved in `canvas/mod.rs` for this step. U8 of the chain extracts
them to `crates/flui-painting/tests/canvas_state.rs` and
`canvas_drawing.rs` etc.

Public API unchanged. External callers (`flui-rendering::context::canvas`,
`flui-engine`, `flui-layer::layer::canvas`) compile without changes.
The `pub mod canvas;` declaration in `lib.rs` resolves to the new
directory automatically.

Verification:
- `cargo check -p flui-painting` clean.
- `cargo test -p flui-painting --lib` 64 passing (matches
  pre-refactor count).
- `cargo test -p flui-painting --tests` 9 passing (integration tests
  unaffected by the structural split).
- `cargo check --workspace` clean.
- `bash scripts/port-check.sh` 6 triggers ok.

Net unsafe delta: 0 (crate is #[forbid(unsafe_code)]; no unsafe
blocks introduced; none removed).
Net LOC delta in production code: -1,118 LOC across the split files
vs. the original god module (duplicate docstring noise + verbose
intermediate boilerplate removed).

Mythos chain step 4 of 14. Next: U5 splits display_list.rs
(2,434 LOC) into a `display_list/` directory.
